feat: Add new components and state management for UI showcase
- Introduced `AppState` to manage application-wide state including performance metrics, button demo data, and text editor states.
- Created `ButtonDemoData` to track button click counts and implemented interactive button components.
- Developed `component_showcase` to organize and display various UI components including buttons, text editors, and performance metrics.
- Added `content_section` for structured layout with headers and text columns.
- Implemented layout examples for outlined and transparent surfaces.
- Created `PerformanceMetrics` to monitor and display frames per second (FPS).
- Added text editor components with customizable selection colors.
- Introduced utility function `create_spacer` for consistent spacing in layouts.
- Enhanced existing components with improved handling of child items and layout constraints.

+++ tessera_basic_components/src/text_edit_core.rs
@@ -325,8 +325,7 @@ pub fn text_edit_core(state: Arc<RwLock<TextEditorState>>) {
 
             // Handle cursor positioning (cursor comes after selection rects)
             if let Some(cursor_pos_raw) = state_clone.read().editor.cursor_position() {
-                let cursor_pos =
-                    PxPosition::new(Px(cursor_pos_raw.0 as i32), Px(cursor_pos_raw.1 as i32));
+                let cursor_pos = PxPosition::new(Px(cursor_pos_raw.0), Px(cursor_pos_raw.1));
                 let cursor_node_index = selection_rects_len;
                 if let Some(cursor_node_id) = input.children_ids.get(cursor_node_index).copied() {
                     let _ = measure_node(
